Moved files to seperate folders model / controller
