+++ app/src/main/java/kline/micah/itsrainingduhitsoregon/controller/MainActivity.java
@@ -1,11 +1,11 @@
-package kline.micah.itsrainingduhitsoregon;
+package kline.micah.itsrainingduhitsoregon.controller;
 
+import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuItem;
+
+import kline.micah.itsrainingduhitsoregon.R;
 
 public abstract class MainActivity extends AppCompatActivity {
