+++ app/src/main/java/kline/micah/itsrainingduhitsoregon/controller/WeatherListActivity.java
@@ -1,10 +1,6 @@
-package kline.micah.itsrainingduhitsoregon;
+package kline.micah.itsrainingduhitsoregon.controller;
 
 import android.support.v4.app.Fragment;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuItem;
 
 public class WeatherListActivity extends MainActivity {
