+++ app/src/main/java/kline/micah/itsrainingduhitsoregon/model/Weather.java
@@ -1,4 +1,4 @@
-package kline.micah.itsrainingduhitsoregon;
+package kline.micah.itsrainingduhitsoregon.model;
 
 import java.util.UUID;
