--- app/src/main/res/layout/activity_weather_details.xml
@@ -1,7 +1,0 @@
-<fragment xmlns:android="http://schemas.android.com/apk/res/android"
-          xmlns:tools="http://schemas.android.com/tools"
-          android:id="@+id/fragment"
-          android:name="kline.micah.itsrainingduhitsoregon.WeatherDetailsFragment"
-          tools:layout="@layout/fragment_weather_details"
-          android:layout_width="match_parent"
-          android:layout_height="match_parent"/>
