+++ app/src/main/res/layout/activity_weather_list.xml
@@ -1,7 +1,7 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
           xmlns:tools="http://schemas.android.com/tools"
           android:id="@+id/fragment"
-          android:name="kline.micah.itsrainingduhitsoregon.WeatherListFragment"
+    android:name="kline.micah.itsrainingduhitsoregon.controller.WeatherListFragment"
           tools:layout="@layout/fragment_weather_list"
           android:layout_width="match_parent"
           android:layout_height="match_parent"/>
