+++ app/src/main/res/layout/fragment_weather_details.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="kline.micah.itsrainingduhitsoregon.WeatherDetailsFragment">
+    tools:context="kline.micah.itsrainingduhitsoregon.controller.WeatherDetailsFragment">
 
     <TextView
         android:id="@+id/highTemperature"
