+++ app/src/main/res/layout/fragment_weather_list.xml
@@ -6,7 +6,7 @@
              android:paddingLeft="@dimen/activity_horizontal_margin"
              android:paddingRight="@dimen/activity_horizontal_margin"
              android:paddingTop="@dimen/activity_vertical_margin"
-             tools:context="kline.micah.itsrainingduhitsoregon.WeatherListFragment"
+    tools:context="kline.micah.itsrainingduhitsoregon.controller.WeatherListFragment"
 
     >
