+++ app/src/main/res/menu/menu_weather_list.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto"
       xmlns:tools="http://schemas.android.com/tools"
-      tools:context="kline.micah.itsrainingduhitsoregon.WeatherListActivity">
+    tools:context="kline.micah.itsrainingduhitsoregon.controller.WeatherListActivity">
     <item android:id="@+id/action_settings"
           android:title="@string/action_settings"
           android:orderInCategory="100"
